+++ doc/source/tutorials/connected_components/assets/connected_components.py
@@ -2,22 +2,21 @@ import igraph as ig
 import matplotlib.pyplot as plt
 import random
 
-# Generate a random geometric graph with random vertex sizes 
+# Generate a random geometric graph with random vertex sizes
 random.seed(0)
 g = ig.Graph.GRG(50, 0.15)
 
 # Cluster graph into weakly connected components
-clusters = g.clusters(mode='weak')
-nclusters = len(clusters)
+components = g.connected_components(mode='weak')
 
 # Visualise different components
 fig, ax = plt.subplots()
 ig.plot(
-    clusters,
+    components,
     target=ax,
     palette=ig.RainbowPalette(),
     vertex_size=0.05,
-    vertex_color=list(map(int, ig.rescale(clusters.membership, (0, 200), clamp=True))),
+    vertex_color=list(map(int, ig.rescale(components.membership, (0, 200), clamp=True))),
     edge_width=0.7
 )