+++ tests/test_games.py
@@ -25,7 +25,7 @@ class GameTests(unittest.TestCase):
     def testPreference(self):
         g = Graph.Preference(100, [1, 1], [[1, 0], [0, 1]])
         self.assertTrue(isinstance(g, Graph))
-        self.assertEqual(len(g.clusters()), 2)
+        self.assertEqual(len(g.connected_components()), 2)
 
         g = Graph.Preference(100, [1, 1], [[1, 0], [0, 1]], attribute="type")
         l = g.vs.get_attribute_values("type")
@@ -34,7 +34,7 @@ class GameTests(unittest.TestCase):
     def testAsymmetricPreference(self):
         g = Graph.Asymmetric_Preference(100, [[0, 1], [1, 0]], [[0, 1], [1, 0]])
         self.assertTrue(isinstance(g, Graph))
-        self.assertEqual(len(g.clusters()), 2)
+        self.assertEqual(len(g.connected_components()), 2)
 
         g = Graph.Asymmetric_Preference(
             100, [[0, 1], [1, 0]], [[1, 0], [0, 1]], attribute="type"
@@ -46,7 +46,7 @@ class GameTests(unittest.TestCase):
 
         g = Graph.Asymmetric_Preference(100, [[0, 1], [1, 0]], [[1, 0], [0, 1]])
         self.assertTrue(isinstance(g, Graph))
-        self.assertEqual(len(g.clusters()), 1)
+        self.assertEqual(len(g.connected_components()), 1)
 
     def testTreeGame(self):
         # Prufer algorithm