+++ tests/test_operators.py
@@ -45,7 +45,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 9
             and g.ecount() == 9
-            and g.clusters().membership == [0, 0, 0, 1, 1, 1, 2, 2, 2]
+            and g.connected_components().membership == [0, 0, 0, 1, 1, 1, 2, 2, 2]
         )
 
     def testDifference(self):
@@ -263,7 +263,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 5
             and g.ecount() == 3
-            and g.clusters().membership == [0, 0, 0, 1, 2]
+            and g.connected_components().membership == [0, 0, 0, 1, 2]
         )
 
         # Adding a vertex by name
@@ -271,7 +271,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 6
             and g.ecount() == 3
-            and g.clusters().membership == [0, 0, 0, 1, 2, 3]
+            and g.connected_components().membership == [0, 0, 0, 1, 2, 3]
         )
 
         # Adding a single edge
@@ -279,13 +279,13 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 6
             and g.ecount() == 4
-            and g.clusters().membership == [0, 0, 0, 0, 1, 2]
+            and g.connected_components().membership == [0, 0, 0, 0, 1, 2]
         )
 
         # Adding two edges
         g += [(3, 4), (2, 4), (4, 5)]
         self.assertTrue(
-            g.vcount() == 6 and g.ecount() == 7 and g.clusters().membership == [0] * 6
+            g.vcount() == 6 and g.ecount() == 7 and g.connected_components().membership == [0] * 6
         )
 
         # Adding two more vertices
@@ -304,7 +304,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 11
             and g.ecount() == 10
-            and g.clusters().membership == [0, 0, 0, 0, 0, 0, 1, 2, 3, 3, 3]
+            and g.connected_components().membership == [0, 0, 0, 0, 0, 0, 1, 2, 3, 3, 3]
         )
 
         # Adding two graphs
@@ -312,7 +312,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 16
             and g.ecount() == 14
-            and g.clusters().membership
+            and g.connected_components().membership
             == [0, 0, 0, 0, 0, 0, 1, 2, 3, 3, 3, 4, 4, 4, 5, 5]
         )
 
@@ -324,7 +324,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 5
             and g.ecount() == 3
-            and g.clusters().membership == [0, 0, 0, 1, 2]
+            and g.connected_components().membership == [0, 0, 0, 1, 2]
         )
         g0 = g
 
@@ -333,7 +333,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 6
             and g.ecount() == 3
-            and g.clusters().membership == [0, 0, 0, 1, 2, 3]
+            and g.connected_components().membership == [0, 0, 0, 1, 2, 3]
         )
         g0 = g
 
@@ -342,14 +342,14 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 6
             and g.ecount() == 4
-            and g.clusters().membership == [0, 0, 0, 0, 1, 2]
+            and g.connected_components().membership == [0, 0, 0, 0, 1, 2]
         )
         g0 = g
 
         # Adding two edges
         g = g0 + [(3, 4), (2, 4), (4, 5)]
         self.assertTrue(
-            g.vcount() == 6 and g.ecount() == 7 and g.clusters().membership == [0] * 6
+            g.vcount() == 6 and g.ecount() == 7 and g.connected_components().membership == [0] * 6
         )
         g0 = g
 
@@ -358,7 +358,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 9
             and g.ecount() == 10
-            and g.clusters().membership == [0, 0, 0, 0, 0, 0, 1, 1, 1]
+            and g.connected_components().membership == [0, 0, 0, 0, 0, 0, 1, 1, 1]
         )
 
     def testInPlaceSubtraction(self):
@@ -370,7 +370,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 7
             and g.ecount() == 21
-            and g.clusters().membership == [0, 0, 0, 0, 0, 0, 0]
+            and g.connected_components().membership == [0, 0, 0, 0, 0, 0, 0]
         )
 
         # Deleting a vertex
@@ -378,7 +378,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 6
             and g.ecount() == 15
-            and g.clusters().membership == [0, 0, 0, 0, 0, 0]
+            and g.connected_components().membership == [0, 0, 0, 0, 0, 0]
         )
 
         # Deleting two vertices
@@ -386,7 +386,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 4
             and g.ecount() == 6
-            and g.clusters().membership == [0, 0, 0, 0]
+            and g.connected_components().membership == [0, 0, 0, 0]
         )
 
         # Deleting an edge
@@ -394,7 +394,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 4
             and g.ecount() == 5
-            and g.clusters().membership == [0, 0, 0, 0]
+            and g.connected_components().membership == [0, 0, 0, 0]
         )
 
         # Deleting three more edges
@@ -402,7 +402,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 4
             and g.ecount() == 2
-            and g.clusters().membership == [0, 0, 1, 1]
+            and g.connected_components().membership == [0, 0, 1, 1]
         )
 
         # Did we really use the original graph so far?
@@ -414,7 +414,7 @@ class OperatorTests(unittest.TestCase):
         self.assertTrue(
             g.vcount() == 4
             and g.ecount() == 1
-            and g.clusters().membership == [0, 1, 2, 2]
+            and g.connected_components().membership == [0, 1, 2, 2]
         )
 
     def testNonzero(self):